requirements.txt
flask
flask-dance
flask-login
flask-socketio
flask-sqlalchemy
gunicorn
pyjwt